requirements.txt
aiofiles==0.6.0
asgiref==3.8.1
channels==4.2.0
click==7.1.2
Django==4.2.19
h11==0.14.0
itsdangerous==1.1.0
MarkupSafe==1.1.1
otree==5.10.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-multipart==0.0.5
six==1.17.0
SQLAlchemy==1.3.22
sqlparse==0.5.3
starlette==0.14.1
typing_extensions==4.12.2
uvicorn==0.13.4
websockets==10.1
WTForms==2.3.3
WTForms-SQLAlchemy==0.2
